thông báo của nút xóa đoàn trong quản lý đoàn

+++ QuanLyTourDuLich/src/com/myclass/gui/Application.java
@@ -2256,11 +2256,21 @@ public class Application extends JFrame {
 		btnDoan_Xoa.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				int selectedRow = tblDoan.getSelectedRow();
-				String maDoan = (String) tblDoan.getValueAt(selectedRow, 0);
-				
-				doanBUS.deleteById(maDoan);
-				doanTblModel.removeRow(selectedRow);
-				DoanBUS.listDoanDTO.remove(selectedRow);
+				if(selectedRow>=0) {
+					int result=JOptionPane.showConfirmDialog(null,"Bạn có chắc muốn xóa thông tin này?","thông báo",JOptionPane.YES_NO_OPTION); 
+					if(result==JOptionPane.YES_OPTION) {
+					String maDoan = (String) tblDoan.getValueAt(selectedRow, 0);
+					doanBUS.deleteById(maDoan);
+					doanTblModel.removeRow(selectedRow);
+					DoanBUS.listDoanDTO.remove(selectedRow);
+					}
+					else if(result==JOptionPane.NO_OPTION) {
+						JOptionPane.showMessageDialog(null,"Thông tin này chưa được xóa!");
+					}
+				}
+				if(selectedRow<0 ) {
+					JOptionPane.showMessageDialog(cardQuanLyTaiKhoan, "Bạn chưa chọn trường dữ liệu nào cả");
+				}
 			}
 		});
 		btnDoan_Xoa.setFont(new Font("Tahoma", Font.PLAIN, 16));
